[wip codegen rewrite] fix marking of varinfo.value when type_is_ghost and recursive constant struct creation
programs should never write to the .typ field directly, since it may
miss that the type is actually ghost, or should be moved to the stack

the recursive case of constant struct to llvm Value* was broken by the
latest changes to move structs to the stack. this fixes that and adds a
test case for it.